+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file as of v2.1.0.
 
 ## [Unreleased]
+### Added
+- Added per page SEO config options
+
 ### Changed
+- Alphabetized controllers.php file
+- Required page numbers to be numeric on blog
 - Updated Readme
 
 ## [2.1.3] - 2018-02-13